Q3-Q4/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
disable-warnings==0.10
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
zipp==3.19.2
